+++ main/library/generators/refinegan.py
@@ -163,11 +163,11 @@ class SineGen(nn.Module):
             for idx in np.arange(self.harmonic_num):
                 f0_buf[:, :, idx + 1] = f0_buf[:, :, 0] * (idx + 2)
 
-            sine_waves = self._f02sine(f0_buf) * self.sine_amp
+            sine_waves = self._f02sine(f0_buf.float()) * self.sine_amp
             uv = self._f02uv(f0)
             sine_waves = sine_waves * uv + ((uv * self.noise_std + (1 - uv) * self.sine_amp / 3) * torch.randn_like(sine_waves))
 
-        return self.merge(sine_waves)
+        return self.merge(sine_waves.to(f0.dtype))
     
 class RefineGANGenerator(nn.Module):
     def __init__(
@@ -245,14 +245,14 @@ class RefineGANGenerator(nn.Module):
             downs.append(x)
 
             x = torchaudio.functional.resample(
-                x.contiguous(), 
+                x.float().contiguous(), 
                 orig_freq=int(f0_size * old_size), 
                 new_freq=int(f0_size * new_size), 
                 lowpass_filter_width=64, 
                 rolloff=0.9475937167399596, 
                 resampling_method="sinc_interp_kaiser", 
                 beta=14.769656459379492
-            )
+            ).to(x.dtype)
             x = block(x)
 
         mel = self.mel_conv(mel)
